Git Title: procurve_Config_pull.py
Git Description:
Added a new argument to the argparser function
if -p is used with any value, the user will be prompted to enter the
password on the command line.

If -p is not used, an environment variable "cyberARK" must be set.

If -p isn't used and cyberARK isnt't set the script will print out
instructions to the command line and exit.