Git Title: procurve_Config_pull.py
Git Description:
Deleted the old send_command_from_file and rewrote to use send_command. Wrote a loop to capture the commands and format a nice output. This eliminates the need for the "end" command to exit configuration mode. Makes the script truly a "discovery" script that can't modify the configuration.

        output_show_str: str = ""
        for command in show_commands:
            output_show = net_connect.send_command(
                command, strip_command=False, delay_factor=1
            )
            output_show_str = output_show_str + "\n" + "!---" + "\n" + output_show