Git Title: procurve_Config_pull.py
Git Description:
A customer had a version of firmware that sent "Note: * indicates values dynamically overridden by RADIUS." as part of the "show interfaces brief" command. This caused the textfsm template to fail and crash. I removed line 17 of the template "^. -> Error" to make it work.

But I had to create a custom template and update the "send_command" command

        template_path = os.getcwd()
        template_file = os.path.join(template_path, "sh_int_br.textfsm")
        print(f"collecting show interfaces brief for {hostname}")
        output_show_int_br = net_connect.send_command(
            "show interfaces brief",
            strip_command=True,
            use_textfsm=True,
            textfsm_template=template_file,
        )

I tested the script on my home lab which doesn't send the radius statement and it continued to work properly.